Define UpdateUserRequestDto
This commit introduces the `UpdateUserRequestDto` data class.

This DTO is used for updating a user's profile information. It allows for partial updates, meaning that only the fields that are being changed need to be populated. The backend will handle null values by keeping the existing data for those fields.

The DTO includes the following properties:
- `firstName`
- `lastName`
- `phoneNumber`
- `gender`
- `dateOfBirth` (as an ISO 8601 date string)
- `profilePictureUrl`

All properties are nullable to support partial updates.