Define InAppAttachmentType enum
This commit introduces the `InAppAttachmentType` enum, which defines the types of structured data entities that can be attached within the application.

This enum is intended for referencing other data within the system, as opposed to external file attachments.

Current types include:
- `CHAPTER`
- `COURSE`
- `EXERCISE`
- `EVALUATION`
- `USER_PROFILE`
- `SCHOOL_ANNOUNCEMENT`

A TODO comment is included to remind to add more types as needed in the future.