Refactor: Rename `usernameOrEmail` to `username` in `LoginRequestDto`
The `usernameOrEmail` field in the `LoginRequestDto` data class has been renamed to `username`.

This change simplifies the field name while still accommodating either a username or an email address as the user's login identifier, depending on the backend configuration. The corresponding KDoc has also been updated to reflect this change.